+++ src/torch_pointcloud/layers/__init__.py
@@ -10,17 +10,17 @@ from ._modules import (
 from .act import create_act
 from .affine import Affine, affine
 from .anchors import (
-    AnchorHeadMulti,
+    AnchorGroupHead,
+    AnchorHead,
     AnchorHeadMultiOutput,
     AnchorHeadOutput,
-    AnchorHeadSingle,
     AnchorTargets,
-    MultiGroupSingleHead,
+    MultiGroupAnchorHead,
     assign_anchor_targets,
     generate_anchors,
     separate_branch,
 )
-from .bev_backbone import BaseBEVBackbone, BaseBEVResBackbone, BasicBlock2d
+from .bev_backbone import BEVBackbone, BEVResidualBackbone, ResidualBlock2d
 from .conv2d_blocks import Conv2dBlock
 from .conv3d_blocks import Conv3dBlock
 from .dropouts import (
@@ -35,7 +35,7 @@ from .geometric_affine import GeometricAffineConv, NormalizeType
 from .grid_pool import GridPool
 from .linear_blocks import LinearBlock
 from .norms import create_norm
-from .octree_attention import RPE, OctreeAttention, OctreeT
+from .octree_attention import OctreeAttention, OctreeRelativePositionEncoding, OctreeT
 from .octree_blocks import OctreeConvBlock, OctreeDeconvBlock
 from .pdnorm import PDNorm
 from .point_patch_embed import PointPatchEmbed
@@ -51,8 +51,6 @@ from .pointnet2_blocks import (
     PointNet2FeaturePropagation,
     PointNet2GlobalSetAbstraction,
     PointNet2SetAbstraction,
-    ensure_msg_list,
-    ensure_msg_list_size,
 )
 from .pointnext_blocks import (
     PointNeXtConv,
@@ -93,6 +91,6 @@ from .spconv_blocks import (
 )
 from .tnet import DynamicTNet, TNet
 from .transformer import Attention, TransformerBlock
-from .vfe import DynamicMeanVFE, PFNLayer
+from .vfe import DynamicMeanVFE, PillarFeatureLayer
 from .view import View
 from .xconv import XConv