+++ src/torch_pointcloud/layers/pointconv_sa.py
@@ -10,7 +10,7 @@ from torch_geometric.typing import OptTensor
 
 from torch_pointcloud.utils.cluster import knn
 from torch_pointcloud.utils.conversion import ensure_list
-from torch_pointcloud.utils.neighbors import gaussian_kernel_density
+from torch_pointcloud.utils.density import gaussian_kernel_density
 
 from .linear_blocks import LinearBlock
 from .pointconv import PointConv, PointConvDensity