+++ src/torch_pointcloud/models/pointnet2.py
@@ -16,12 +16,8 @@ from torch_pointcloud.datasets.modelnet import MODELNET40_CLASSES
 from torch_pointcloud.datasets.s3dis import S3DIS_CLASSES
 from torch_pointcloud.datasets.scanobjectnn import SCANOBJECTNN_CLASSES
 from torch_pointcloud.layers import PoolLike, create_pool
-from torch_pointcloud.layers.pointnet2_blocks import (
-    PointNet2FeaturePropagation,
-    PointNet2SetAbstraction,
-    ensure_msg_list,
-)
-from torch_pointcloud.utils.conversion import ensure_list, ensure_tuple, ensure_tuple_size
+from torch_pointcloud.layers.pointnet2_blocks import PointNet2FeaturePropagation, PointNet2SetAbstraction
+from torch_pointcloud.utils.conversion import ensure_list, ensure_msg_list, ensure_tuple, ensure_tuple_size
 from torch_pointcloud.utils.data import DataKeys
 from torch_pointcloud.utils.types import AggrType, FeaturesDict, OptTensor